requirements.txt
torch==2.7.1
torchvision==0.22.1
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2
opencv-python-headless==4.10.0.82
pandas
Pillow
wandb==0.22.2
kaggle==1.7.4.5
kagglehub==0.3.13
accelerate
hydra-core
omegaconf
tqdm
dlib